refactor: 인증 API 파일을 auth 도메인 하위로 이동
apis/auth.ts -> feature/auth/api.ts
